module::ModuleStatus: Move to separate header file
